♻️ Rename event handlers to On<EventName>
The composition event handlers (former *Subscriber.cs files, all now
implementing ICompositionEventsHandler<T>) were still using the old
v2 'Subscriber' suffix. NServiceBus already has its own 'Handler'
naming; reusing that here would conflict on conversational read.
Settle on On<EventName> instead: file name + class name match, and
the class declaration reads directly as a handler-style hook for
the event - OnCartLoaded, OnProductLoaded, OnDeliveryOptionsLoaded.

Catalog.ServiceComposition/Products/OnSummaryLoaded.cs also had its
namespace wrong (Finance.ServiceComposition.Products); fixed.

No external references touched - the renamed classes are only
discovered via the ICompositionEventsHandler<T> interface and the
assembly scanner.
