+++ src/Finance.ServiceComposition/Checkout/OnDeliveryOptionsLoaded.cs
@@ -6,7 +6,7 @@ using Shipping.ServiceComposition.Events;
 
 namespace Finance.ServiceComposition.Checkout;
 
-public class DeliveryOptionsLoadedSubscriber(FinanceDbContext dbContext) : ICompositionEventsHandler<DeliveryOptionsLoaded>
+public class OnDeliveryOptionsLoaded(FinanceDbContext dbContext) : ICompositionEventsHandler<DeliveryOptionsLoaded>
 {
     public async Task Handle(DeliveryOptionsLoaded @event, HttpRequest request)
     {
