+++ src/Finance.ServiceComposition/Products/OnProductsLoaded.cs
@@ -1,32 +1,25 @@
 using Catalog.ServiceComposition.Events;
 using Finance.Data;
-using Finance.Data.Models;
 using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using ServiceComposer.AspNetCore;
 
-namespace Finance.ServiceComposition.Cart;
+namespace Finance.ServiceComposition.Products;
 
-public class CartLoadedSubscriber(FinanceDbContext dbContext) : ICompositionEventsHandler<CartLoaded>
+class OnProductsLoaded(FinanceDbContext dbContext) : ICompositionEventsHandler<ProductsLoaded>
 {
-    public async Task Handle(CartLoaded @event, HttpRequest request)
+    public async Task Handle(ProductsLoaded @event, HttpRequest request)
     {
-        var productIds = @event.OrderedProducts.Keys.ToList();
+        var productIds = @event.Products.Keys.ToList();
         var resultSet = await dbContext.Products
             .Where(s => productIds.Contains(s.ProductId))
             .ToListAsync(request.HttpContext.RequestAborted);
 
-        decimal totalPrice = 0;
-        foreach (var product in @event.OrderedProducts)
+        foreach (var product in @event.Products)
         {
             var matchingProduct = resultSet.Single(s => s.ProductId == product.Key);
             product.Value.Price = matchingProduct.Price;
             product.Value.Discount = matchingProduct.Discount;
-
-            totalPrice += matchingProduct.EffectivePrice() * (int)product.Value.Quantity;
         }
-
-        var vm = request.GetComposedResponseModel();
-        vm.TotalCartPrice = totalPrice;
     }
 }
